+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 * **Changes:**
   * Renamed `ElasticsearchEx.Api.Document.get_document/1` to `ElasticsearchEx.Api.Document.get/1`
   * Replaced `ElasticsearchEx.Api.Document.get_source/1` by `ElasticsearchEx.Api.Document.get/1` with `source_only: true` option
+  * Renamed `ElasticsearchEx.Api.Document.document_exists?/1` to `ElasticsearchEx.Api.Document.exists?/1`
+  * Replaced `ElasticsearchEx.Api.Document.source_exists?/1` by `ElasticsearchEx.Api.Document.exists?/1` with `source_only: true` option
 
 ## v0.6.2 (2024-04-01)